Move forkserver into a dedicated binary
PiperOrigin-RevId: 242637894
