+++ sandboxed_api/sandbox2/global_forkclient.cc
@@ -16,8 +16,10 @@
 
 #include "sandboxed_api/sandbox2/global_forkclient.h"
 
+#include <fcntl.h>
 #include <sys/prctl.h>
 #include <sys/socket.h>
+#include <syscall.h>
 #include <unistd.h>
 
 #include <csignal>
@@ -25,8 +27,10 @@
 
 #include "absl/base/attributes.h"
 #include "absl/strings/str_cat.h"
+#include "sandboxed_api/embed_file.h"
 #include "sandboxed_api/sandbox2/comms.h"
 #include "sandboxed_api/sandbox2/forkserver.h"
+#include "sandboxed_api/sandbox2/forkserver_bin_embed.h"
 #include "sandboxed_api/sandbox2/sanitizer.h"
 #include "sandboxed_api/sandbox2/util/strerror.h"
 #include "sandboxed_api/util/raw_logging.h"
@@ -84,43 +88,20 @@ static void StartGlobalForkServer() {
     return;
   }
 
-  // Make sure the logs go stderr.
-  google::LogToStderr();
+  // Move the comms FD to the proper, expected FD number.
+  // The new FD will not be CLOEXEC, which is what we want.
+  dup2(sv[0], Comms::kSandbox2ClientCommsFD);
 
-  // Close all non-essential FDs to keep newly opened FD numbers consistent.
-  sandbox2::sanitizer::CloseAllFDsExcept({0, 1, 2, sv[0]});
-
-  // Make the process' name easily recognizable with ps/pstree.
-  if (prctl(PR_SET_NAME, "S2-FORK-SERV", 0, 0, 0) != 0) {
-    SAPI_RAW_PLOG(WARNING, "prctl(PR_SET_NAME, 'S2-FORK-SERV')");
+  int exec_fd = sapi::EmbedFile::GetEmbedFileSingleton()->GetFdForFileToc(
+      forkserver_bin_embed_create());
+  if (exec_fd < 0) {
+    SAPI_RAW_LOG(FATAL, "Getting FD for init binary failed");
   }
 
-  // Don't react (with stack-tracing) to SIGTERM's sent from other processes
-  // (e.g. from the borglet or SubProcess). This ForkServer should go down if
-  // the parent goes down (or if the GlobalForkServerComms is closed), which is
-  // assured by prctl(PR_SET_PDEATHSIG, SIGKILL) being called in the
-  // ForkServer::Initialize(). We don't want to change behavior of non-global
-  // ForkServers, hence it's called here and not in the
-  // ForkServer::Initialize().
-  struct sigaction sa;
-  sa.sa_handler = SIG_IGN;
-  sa.sa_flags = 0;
-  sigemptyset(&sa.sa_mask);
-  if (sigaction(SIGTERM, &sa, nullptr) == -1) {
-    SAPI_RAW_PLOG(WARNING, "sigaction(SIGTERM, sa_handler=SIG_IGN)");
-  }
-
-  Comms comms(sv[0]);
-  ForkServer fork_server(&comms);
-
-  while (true) {
-    pid_t child_pid = fork_server.ServeRequest();
-    if (!child_pid) {
-      // FORKSERVER_FORK sent to the global forkserver. This case does not make
-      // sense, we thus kill the process here.
-      exit(0);
-    }
-  }
+  char* const args[] = {strdup("S2-FORK-SERV"), nullptr};
+  char* const envp[] = {nullptr};
+  syscall(__NR_execveat, exec_fd, "", args, envp, AT_EMPTY_PATH);
+  SAPI_RAW_PCHECK(false, "Could not launch forkserver binary");
 }
 
 }  // namespace sandbox2
